# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Leave Management System API
  description: API documentation for the Leave Management System with role-based access control
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:8081
    description: Alternative local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginCredentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterCredentials:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
...
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCredentials'
      responses:
        201:
          description: User registered successfully
        400:
          description: Invalid input or user already exists
  /api/auth/login:
    post:
      summary: Login user and get access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        200:
          description: Login successful, returns access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        401:
          description: Invalid credentials
  /api/auth/validate-token:
    post:
      summary: Validate access token
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        200:
          description: Token is valid
        401:
          description: Invalid or expired token

  # Leave Request Routes
...